Dialer: fix sub name not updated in msim call log filter
Original codes use Setting.System.getString() to get names,
but this way is obsolete and will return null, will use SIM1
and SIM2 as names, so not updated after user changes SIM names
in settings.

We'll get the sub names by phone account label, which is right
sub display name, and this is corresponding to names in setting.

Change-Id: Ic59e91c0f95e59408b79141a93e8abe2b7531f38
CRs-Fixed: 731519
Signed-off-by: houst0nn <houstonn@gmail.com>